requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
flask>=3.0.0,<4
python-dotenv>=1.0.1
requests>=2.31.0
torch==2.3.1+cpu
ollama>=0.3.1
PyPDF2>=3.0.1
numpy>=1.26.4
